refactor: change how stdout/stderr are handled by backends.
Currently, each backend is returning stdout and stderr as part of the
`std::process::Output` returned by task execution.

For the Docker backend, this was correctly the stdout and stderr of the
container.

For TES, however, the returned stdout and stderr were from TES API fields that
don't necessary contain relevant information. According to the TES API
documentation, the only way to consistently get stdout and stderr is to specify
their container paths for the executor and add outputs for copying them to
storage.

For the generic backend, we were always returning the stdout and stderr of the
last command run, which typically was for the monitoring script.

This commit changes it so only the exit status is returned from task execution;
instead, we now follow the TES model of requiring the `stdout` and `stderr`
fields be used on the execution and add outputs to specify where those files
should be stored as outputs.

This corrects the behavior for the TES backend and also enables the Docker
backend to persist the stdout and stderr streams directly to the file system
rather than always buffering them in memory.

With these changes, the generic backend no longer returns stdout or stderr as
we'll need additional support in the generic backend for running commands that
will persist the task stdout or stderr to a location.

Additionally, builder types have had their fields made public so that the
backends don't have to clone their contents.

The TES backend now has a configuration option for controlling the polling
interval.